Engineering Task #3303 Domain Object and Interceptor for ServiceRequest Audit Logging
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@16725 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ domain/src/main/java/uk/gov/moj/sdt/domain/AbstractDomainObject.java
@@ -48,7 +48,7 @@ public abstract class AbstractDomainObject implements IDomainObject, IVisitable
     /**
      * Primary key.
      */
-    private long id;
+    private int id;
 
     /**
      * Hibernate version number.
@@ -80,13 +80,13 @@ public abstract class AbstractDomainObject implements IDomainObject, IVisitable
     }
 
     @Override
-    public long getId ()
+    public int getId ()
     {
         return id;
     }
 
     @Override
-    public void setId (final long id)
+    public void setId (final int id)
     {
         this.id = id;
     }
